requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
-e file:external/gobotany-app
    # via -r requirements.in
-e file:.
    # via -r requirements.in
annotated-types==0.5.0
    # via pydantic
asgiref==3.7.2
    # via django
async-timeout==4.0.3
    # via redis
beautifulsoup4==4.9.3
    # via gobotany
bleach==3.3.0
    # via gobotany
boto==2.49.0
    # via gobotany
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
click==8.1.6
    # via rq
collectfast==2.2.0
    # via goorchids
confusable-homoglyphs==3.2.0
    # via django-registration
cssselect==1.2.0
    # via gobotany
defusedxml==0.7.1
    # via odfpy
django==3.1.13
    # via
    #   collectfast
    #   django-appconf
    #   django-debug-toolbar
    #   django-extensions
    #   django-haystack
    #   django-registration
    #   django-storages
    #   django-user-accounts
    #   gobotany
django-appconf==1.0.5
    # via
    #   django-imagekit
    #   django-user-accounts
django-debug-toolbar==3.2.1
    # via gobotany
django-extensions==3.1.2
    # via gobotany
django-haystack==3.0
    # via gobotany
django-imagekit==4.0.2
    # via gobotany
django-pylibmc==0.6.1
    # via gobotany
django-registration==3.1.2
    # via gobotany
django-storages==1.7.2
    # via
    #   collectfast
    #   gobotany
django-tinymce==2.7.1
    # via
    #   -r requirements.in
    #   gobotany
django-user-accounts==3.0.2
    # via gobotany
et-xmlfile==1.1.0
    # via openpyxl
gunicorn==20.0.4
    # via gobotany
honcho==1.1.0
    # via goorchids
idna==3.4
    # via requests
inflect==7.0.0
    # via gobotany
lxml==4.6.3
    # via gobotany
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via tablib
packaging==23.1
    # via bleach
pilkit==2.0
    # via django-imagekit
pillow==8.2.0
    # via gobotany
psycopg2==2.9.2
    # via gobotany
pydantic==2.1.1
    # via inflect
pydantic-core==2.4.0
    # via pydantic
pylibmc==1.5.2
    # via
    #   django-pylibmc
    #   gobotany
pysolr==3.9.0
    # via gobotany
python-dateutil==2.8.2
    # via s3cmd
python-magic==0.4.24
    # via
    #   gobotany
    #   s3cmd
python-memcached==1.59
    # via gobotany
pytz==2023.3
    # via
    #   django
    #   django-user-accounts
    #   gobotany
pyyaml==6.0.1
    # via tablib
redis==5.0.0
    # via
    #   goorchids
    #   rq
requests==2.31.0
    # via
    #   gobotany
    #   pysolr
rq==1.15.1
    # via goorchids
s3cmd==2.2.0
    # via gobotany
setuptools-scm==4.1.2
    # via -r requirements.in
six==1.16.0
    # via
    #   bleach
    #   django-imagekit
    #   python-dateutil
    #   python-memcached
soupsieve==2.4.1
    # via beautifulsoup4
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
tablib==0.12.1
    # via gobotany
typing-extensions==4.7.1
    # via
    #   asgiref
    #   collectfast
    #   inflect
    #   pydantic
    #   pydantic-core
unicodecsv==0.14.1
    # via tablib
urllib3==2.0.4
    # via requests
webencodings==0.5.1
    # via bleach
xlrd==2.0.1
    # via
    #   gobotany
    #   tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
